refactor: organize forget password screen to reset password package and rename it
