+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/resetPassword/phoneEntry/ForgetPasswordPhoneEntryScreenInteractionListener.kt
@@ -1,9 +1,9 @@
-package net.thechance.mena.identity.presentation.screen.forgetPassword
+package net.thechance.mena.identity.presentation.screen.resetPassword.phoneEntry
 
 import net.thechance.mena.identity.presentation.base.BaseInteractionListener
 import net.thechance.mena.identity.presentation.screen.countryPicker.menaCountries.MenaCountry
 
-interface ForgetPasswordScreenInteractionListener : BaseInteractionListener{
+interface ForgetPasswordPhoneEntryScreenInteractionListener : BaseInteractionListener{
     fun onSelectCountryItem(country: MenaCountry)
     fun onDismissBottomSheet()
     fun onClickContinue()
